requirements.txt
Flask
azure-cosmos
python-dotenv